offset als double
git-svn-id: svn://ntoserver/Datenverteiler/VRZ3/trunk/de.bsvrz.sys.funclib.bitctrl@6697 40c0be50-5d29-0410-901d-fb59edb5cafc

+++ src/de/bsvrz/sys/funclib/bitctrl/modell/verkehr/objekte/StrassenTeilSegment.java
@@ -433,12 +433,12 @@ public class StrassenTeilSegment extends StoerfallIndikator implements LinieXY {
 	 * @return der Offset des zugeh&ouml;rigen Stra&szlig;ensegmentes, bei dem
 	 *         das Teilsegment beginnt
 	 */
-	public long getSegmentOffsetAnfang() {
+	public double getSegmentOffsetAnfang() {
 		if (strassenSegment == null) {
 			getStrassenSegment();
 		}
 
-		long offset = -1;
+		double offset = -1;
 
 		if (strassenSegment != null) {
 			offset = 0;
